smoothing without rounding b to internal leaves and cleanup

+++ vowpalwabbit/pmf_to_pdf.h
@@ -19,9 +19,9 @@ struct reduction
   void learn(example& ec);
 
   ~reduction();
-  std::vector<uint32_t> pdf_lim;
+  std::vector<float> pdf_lim;
   uint32_t num_actions;
-  uint32_t bandwidth;  // radius
+  float bandwidth; // radius
   float min_value;
   float max_value;
   bool first_only;
